Crate kubos_system

KubOS System level APIs

Modules

logger
Common structure and functions for setting up service logging

Structs

Address
A simple address consisting of an IP address and port number

Config
KubOS config used by either Apps or Services. KubOS config files use the TOML format, and can may contain multiple named Categories. Typically each category corresponds to an App or Service name. This allows one config file to store configuration for multiple Apps / Services at a time.

KubosVersions
Information about the version(s) of KubOS installed in the system

UBootVars
A convenience wrapper for fetching UBoot variables used by KubOS

Constants

SERVICE_APP
The name of the KubOS app service that can be used to derive service configuration

SERVICE_TELEMETRY
The name of the KubOS telemetry db service that can be used to dervice service configuration

Statics

DEFAULT_CONFIG_PATH
The default configuration file path

DEFAULT_PATH
The default configuration file path

Functions

initial_deploy
Whether or not the system has been marked as deployed

kubos_versions
Fetch information about the version(s) of KubOS installed in the system